Wrap all big chunks of logic into functions to make running and testing easier.

+++ download_data.py
@@ -1,6 +1,7 @@
 from urllib.request import urlretrieve
 from tqdm import tqdm
 import tarfile
+import os.path
 
 
 # From http://ufldl.stanford.edu/housenumbers/
@@ -20,7 +21,7 @@ extra_32 = "http://ufldl.stanford.edu/housenumbers/extra_32x32.mat"
 cropped_urls = [train_32, test_32]
 
 
-class TqdmUpTo(tqdm):
+class ProgressUpTo(tqdm):
     """Provides `update_to(n)` which uses `tqdm.update(delta_n)`."""
     def update_to(self, b=1, bsize=1, tsize=None):
         """
@@ -36,25 +37,40 @@ class TqdmUpTo(tqdm):
         self.update(b * bsize - self.n)  # will also set self.n = b * bsize
 
 
-def download_tar(url, local_dir_path, local_file_name=None):
+def download_tar(url, local_dir_path, local_file_name=None, overwrite=False):
     local_file_name = local_file_name or url.split('/')[-1]
     local_file_path = '/'.join([local_dir_path, local_file_name])
 
-    with TqdmUpTo(unit='B', unit_scale=True, miniters=1, desc=url.split('/')[-1]) as t:
+    if os.path.exists(local_file_path) and ~overwrite:
+        print('File {} already exists locally at {}, skipping download.'.format(local_file_name, local_file_path))
+        return local_file_name, local_file_path
+
+    print("Downloading file from {}...".format(url))
+    with ProgressUpTo(unit='B', unit_scale=True, miniters=1, desc=url.split('/')[-1]) as t:
         path, message = urlretrieve(url, local_file_path, reporthook=t.update_to)
 
-    return local_file_name, path, message
+    return local_file_name, path
 
 
-def extract_tar(file_path):
-    f = tarfile.open(file_path)
+def extract_tar(file_path, overwrite=False):
     dir_comps = file_path.split('/')
+    dir_name = dir_comps[-1].split('.')[0]
     dir_path = '/'.join(dir_comps[:-1])
+    out_dir_name = '/'.join([dir_path, dir_name])
+
+    if os.path.exists(dir_path) and ~overwrite:
+        print('Extracted information for {} already exists in {}, skipping extraction.'.format(file_path, out_dir_name))
+        return
+
+    print("Extracting {} to {}...".format(file_path, dir_path))
+
+    f = tarfile.open(file_path)
+
     f.extractall(path=dir_path)
 
 
 def download_and_extract_tar(url, local_dir_path, local_file_name=None):
-    name, path, _ = download_tar(url, local_dir_path, local_file_name)
+    name, path = download_tar(url, local_dir_path, local_file_name)
     extract_tar(path)
